Moved Protocol.h to shared folder
