+++ client/wrapper/CMakeLists.txt
@@ -1,4 +1,2 @@
-include_directories(${CMAKE_SOURCE_DIR}/client)
-
 add_subdirectory(Java)
 add_subdirectory(Lua)
