+++ server/CMakeLists.txt
@@ -1,6 +1,13 @@
+set(CMAKE_INTERPROCEDURAL_OPTIMIZATION TRUE)
+
 configure_file(Info.h.in Info.h)
 
-include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR} ${CMAKE_SOURCE_DIR}/minparse/include) # A bit messy
+include_directories( # A bit messy
+    ${CMAKE_CURRENT_BINARY_DIR}
+    ${CMAKE_SOURCE_DIR}/server
+    ${CMAKE_SOURCE_DIR}/shared
+    ${CMAKE_SOURCE_DIR}/minparse/include
+)
 
 # This should probably be improved, the core server could potentially become quite large
 add_executable(Server
@@ -12,5 +19,3 @@ add_executable(Server
     processing/CScheduler.cpp
     processing/modules/translate/Translate.cpp
 )
-
-set_property(TARGET Server PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
