+++ tests/subsystem/CMakeLists.txt
@@ -1,4 +1,7 @@
-include_directories(${CMAKE_SOURCE_DIR}/client)
+include_directories(
+    ${CMAKE_SOURCE_DIR}/client
+    ${CMAKE_SOURCE_DIR}/shared
+)
 
 add_executable(Test_Subsystem Run.cpp)
